move icons to utils
