+++ express/code/blocks/frictionless-quick-action-mobile/frictionless-quick-action-mobile.js
@@ -2,9 +2,9 @@ import {
   createTag,
   getLibs,
   getMobileOperatingSystem,
+  getIconElementDeprecated,
 } from '../../scripts/utils.js';
 import { transformLinkToAnimation } from '../../scripts/utils/media.js';
-import { getIconElementDeprecated } from '../../scripts/utils/icons.js';
 
 let replaceKey; let getConfig;
 let loadScript;
