+++ express/code/blocks/frictionless-quick-action/frictionless-quick-action.js
@@ -2,11 +2,10 @@ import {
   addAnimationToggle,
   transformLinkToAnimation,
 } from '../../scripts/utils/media.js';
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs, getIconElementDeprecated } from '../../scripts/utils.js';
 import { buildFreePlanWidget } from '../../scripts/widgets/free-plan.js';
 import { decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
 import { sendFrictionlessEventToAdobeAnaltics } from '../../scripts/instrument.js';
-import { getIconElementDeprecated } from '../../scripts/utils/icons.js';
 
 let createTag; let getConfig;
 let loadScript; let getMetadata;
