+++ express/code/blocks/how-to-steps/how-to-steps.js
@@ -1,7 +1,6 @@
 /* eslint-disable import/named, import/extensions */
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs, getIconElementDeprecated } from '../../scripts/utils.js';
 import { decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
-import { getIconElementDeprecated } from '../../scripts/utils/icons.js';
 
 let createTag;
