+++ express/code/blocks/pricing-cards-credits/pricing-cards-credits.js
@@ -1,6 +1,5 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs, fixIcons } from '../../scripts/utils.js';
 import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
-import { fixIcons } from '../../scripts/utils/icons.js';
 
 let createTag;
