+++ express/code/blocks/pricing-table/pricing-table.js
@@ -1,9 +1,8 @@
-import { getLibs, yieldToMain } from '../../scripts/utils.js';
+import { getLibs, yieldToMain, fixIcons } from '../../scripts/utils.js';
 import { debounce } from '../../scripts/utils/hofs.js';
 import { decorateButtonsDeprecated, addTempWrapperDeprecated, splitAndAddVariantsWithDash } from '../../scripts/utils/decorate.js';
 import { formatDynamicCartLink } from '../../scripts/utils/pricing.js';
 import { sendEventToAnalytics } from '../../scripts/instrument.js';
-import { fixIcons } from '../../scripts/utils/icons.js';
 
 let createTag; let getConfig;
 let replaceKey;
