+++ express/code/blocks/template-list/template-list.js
@@ -4,12 +4,13 @@ import {
   getLottie,
   lazyLoadLottiePlayer,
   toClassName,
+  fixIcons, getIconElementDeprecated,
 } from '../../scripts/utils.js';
 
 import { addAnimationToggle, createOptimizedPicture, linkImage } from '../../scripts/utils/media.js';
 import { fetchRelevantRows } from '../../scripts/utils/relevant.js';
 
-import { decorateSocialIcons, fixIcons, getIconElementDeprecated } from '../../scripts/utils/icons.js';
+import { decorateSocialIcons } from '../../scripts/utils/icons.js';
 import { addTempWrapperDeprecated, decorateButtonsDeprecated } from '../../scripts/utils/decorate.js';
 
 import { Masonry } from '../../scripts/widgets/masonry.js';
