+++ express/code/blocks/toggle-bar/toggle-bar.js
@@ -1,7 +1,6 @@
-import { getLibs, readBlockConfig } from '../../scripts/utils.js';
+import { getLibs, fixIcons, readBlockConfig } from '../../scripts/utils.js';
 import { addTempWrapperDeprecated } from '../../scripts/utils/decorate.js';
 import { sendEventToAnalytics, textToName } from '../../scripts/instrument.js';
-import { fixIcons } from '../../scripts/utils/icons.js';
 
 let createTag;
